pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
    "httpx>=0.25.0",
    "fastmcp>=0.1.0",
    "mysql-connector-python>=8.0.0",
]
